+++ packages/client/components/BackButton.tsx
@@ -2,7 +2,6 @@ import styled from '@emotion/styled'
 import React from 'react'
 import {Link} from 'react-router-dom'
 import {PALETTE} from '~/styles/paletteV3'
-
 import FlatButton from './FlatButton'
 import Icon from './Icon'