+++ packages/client/components/GitHubScopingSearchFilterToggle.tsx
@@ -1,25 +1,17 @@
-import styled from '@emotion/styled'
 import graphql from 'babel-plugin-relay/macro'
 import React from 'react'
 import {useFragment} from 'react-relay'
 import {MenuPosition} from '../hooks/useCoords'
 import useMenu from '../hooks/useMenu'
-import {PALETTE} from '../styles/paletteV3'
-import {ICON_SIZE} from '../styles/typographyV2'
 import lazyPreload from '../utils/lazyPreload'
 import {GitHubScopingSearchFilterToggle_meeting$key} from '../__generated__/GitHubScopingSearchFilterToggle_meeting.graphql'
-import Icon from './Icon'
-import PlainButton from './PlainButton/PlainButton'
+import FilterButton from './FilterButton'
 
-const FilterIcon = styled(Icon)({
-  color: PALETTE.SLATE_600,
-  fontSize: ICON_SIZE.MD24
-})
-
-const GitHubScopingSearchFilterMenuRoot = lazyPreload(() =>
-  import(
-    /* webpackChunkName: 'GitHubScopingSearchFilterMenuRoot' */ './GitHubScopingSearchFilterMenuRoot'
-  )
+const GitHubScopingSearchFilterMenuRoot = lazyPreload(
+  () =>
+    import(
+      /* webpackChunkName: 'GitHubScopingSearchFilterMenuRoot' */ './GitHubScopingSearchFilterMenuRoot'
+    )
 )
 interface Props {
   meetingRef: GitHubScopingSearchFilterToggle_meeting$key
@@ -44,9 +36,7 @@ const GitHubScopingSearchFilterToggle = (props: Props) => {
   })
   return (
     <>
-      <PlainButton onClick={togglePortal} ref={originRef}>
-        <FilterIcon>filter_list</FilterIcon>
-      </PlainButton>
+      <FilterButton onClick={togglePortal} ref={originRef} />
       {menuPortal(
         <GitHubScopingSearchFilterMenuRoot
           teamId={teamId}