+++ packages/client/components/ParabolScopingSearchFilterToggle.tsx
@@ -1,20 +1,11 @@
 import graphql from 'babel-plugin-relay/macro'
 import React from 'react'
 import {createFragmentContainer} from 'react-relay'
-import {ParabolScopingSearchFilterToggle_meeting} from '../__generated__/ParabolScopingSearchFilterToggle_meeting.graphql'
-import PlainButton from './PlainButton/PlainButton'
-import styled from '@emotion/styled'
-import Icon from './Icon'
-import {PALETTE} from '~/styles/paletteV3'
-import {ICON_SIZE} from '~/styles/typographyV2'
-import useMenu from '~/hooks/useMenu'
 import {MenuPosition} from '~/hooks/useCoords'
+import useMenu from '~/hooks/useMenu'
+import {ParabolScopingSearchFilterToggle_meeting} from '../__generated__/ParabolScopingSearchFilterToggle_meeting.graphql'
+import FilterButton from './FilterButton'
 import ParabolScopingSearchFilterMenu from './ParabolScopingSearchFilterMenu'
-
-const FilterIcon = styled(Icon)({
-  color: PALETTE.SLATE_600,
-  fontSize: ICON_SIZE.MD24
-})
 
 interface Props {
   meeting: ParabolScopingSearchFilterToggle_meeting
@@ -28,9 +19,7 @@ const ParabolScopingSearchFilterToggle = (props: Props) => {
   })
   return (
     <>
-      <PlainButton onClick={togglePortal} ref={originRef}>
-        <FilterIcon>filter_list</FilterIcon>
-      </PlainButton>
+      <FilterButton onClick={togglePortal} ref={originRef} />
       {menuPortal(<ParabolScopingSearchFilterMenu meeting={meeting} menuProps={menuProps} />)}
     </>
   )