+++ packages/client/components/ScopePhaseArea.tsx
@@ -8,12 +8,14 @@ import {Breakpoint} from '~/types/constEnums'
 import {ScopePhaseArea_meeting} from '~/__generated__/ScopePhaseArea_meeting.graphql'
 import {Elevation} from '../styles/elevation'
 import {PALETTE} from '../styles/paletteV3'
+import AzureDevOpsSVG from './AzureDevOpsSVG'
 import GitHubSVG from './GitHubSVG'
 import GitLabSVG from './GitLabSVG'
 import Icon from './Icon'
 import JiraServerSVG from './JiraServerSVG'
 import JiraSVG from './JiraSVG'
 import ParabolLogoSVG from './ParabolLogoSVG'
+import ScopePhaseAreaAzureDevOps from './ScopePhaseAreaAzureDevOps'
 import ScopePhaseAreaGitHub from './ScopePhaseAreaGitHub'
 import ScopePhaseAreaGitLab from './ScopePhaseAreaGitLab'
 import ScopePhaseAreaJira from './ScopePhaseAreaJira'
@@ -70,24 +72,42 @@ const ScopePhaseArea = (props: Props) => {
   const {meeting} = props
   const isDesktop = useBreakpoint(Breakpoint.SIDEBAR_LEFT)
   const {viewerMeetingMember} = meeting
-  if (!viewerMeetingMember) return null
-  const {user, teamMember} = viewerMeetingMember
-  const {featureFlags} = user
-  const gitlabIntegration = teamMember.integrations.gitlab
-  const jiraServerIntegration = teamMember.integrations.jiraServer
+  const featureFlags = viewerMeetingMember?.user.featureFlags
+  const gitlabIntegration = viewerMeetingMember?.teamMember.integrations.gitlab
+  const jiraServerIntegration = viewerMeetingMember?.teamMember.integrations.jiraServer
+  const azureDevOpsIntegration = viewerMeetingMember?.teamMember.integrations.azureDevOps
+
   const isGitLabProviderAvailable = !!(
-    gitlabIntegration.cloudProvider?.clientId || gitlabIntegration.sharedProviders.length
+    gitlabIntegration?.cloudProvider?.clientId || gitlabIntegration?.sharedProviders.length
   )
 
-  const allowGitLab = isGitLabProviderAvailable && featureFlags.gitlab
-  const allowJiraServer = !!jiraServerIntegration.sharedProviders.length
+  const allowAzureDevOps =
+    !!azureDevOpsIntegration?.sharedProviders.length && featureFlags?.azureDevOps
+  const allowGitLab = !!(isGitLabProviderAvailable && featureFlags?.gitlab)
+  const allowJiraServer = !!jiraServerIntegration?.sharedProviders.length
 
   const baseTabs = [
-    {icon: <GitHubSVG />, label: 'GitHub', allow: true},
-    {icon: <JiraSVG />, label: 'Jira', allow: true},
-    {icon: <ParabolLogoSVG />, label: 'Parabol', allow: true},
-    {icon: <JiraServerSVG />, label: 'Jira Server', allow: allowJiraServer},
-    {icon: <GitLabSVG />, label: 'GitLab', allow: allowGitLab}
+    {icon: <GitHubSVG />, label: 'GitHub', allow: true, Component: ScopePhaseAreaGitHub},
+    {icon: <JiraSVG />, label: 'Jira', allow: true, Component: ScopePhaseAreaJira},
+    {
+      icon: <ParabolLogoSVG />,
+      label: 'Parabol',
+      allow: true,
+      Component: ScopePhaseAreaParabolScoping
+    },
+    {
+      icon: <JiraServerSVG />,
+      label: 'Jira Server',
+      allow: allowJiraServer,
+      Component: ScopePhaseAreaJiraServer
+    },
+    {icon: <GitLabSVG />, label: 'GitLab', allow: allowGitLab, Component: ScopePhaseAreaGitLab},
+    {
+      icon: <AzureDevOpsSVG />,
+      label: 'Azure DevOps',
+      allow: allowAzureDevOps,
+      Component: ScopePhaseAreaAzureDevOps
+    }
   ] as const
 
   const tabs = baseTabs.filter(({allow}) => allow)
@@ -96,6 +116,7 @@ const ScopePhaseArea = (props: Props) => {
     const idx = tabs.findIndex((tab) => tab.label === favoriteService)
     return idx === -1 ? 1 : idx
   })
+
   const isTabActive = (label: typeof baseTabs[number]['label']) => {
     return activeIdx === tabs.findIndex((tab) => tab.label === label)
   }
@@ -113,7 +134,7 @@ const ScopePhaseArea = (props: Props) => {
     selectIdx(idx)
   }
 
-  const goToParabol = () => {
+  const gotoParabol = () => {
     setActiveIdx(2)
   }
 
@@ -140,37 +161,16 @@ const ScopePhaseArea = (props: Props) => {
         containerStyle={containerStyle}
         style={innerStyle}
       >
-        <TabContents>
-          <ScopePhaseAreaGitHub
-            isActive={isTabActive('GitHub')}
-            gotoParabol={goToParabol}
-            meetingRef={meeting}
-          />
-        </TabContents>
-        <TabContents>
-          <ScopePhaseAreaJira
-            isActive={isTabActive('Jira')}
-            gotoParabol={goToParabol}
-            meeting={meeting}
-          />
-        </TabContents>
-        <TabContents>
-          <ScopePhaseAreaParabolScoping isActive={isTabActive('Parabol')} meeting={meeting} />
-        </TabContents>
-        <TabContents>
-          <ScopePhaseAreaJiraServer
-            isActive={isTabActive('Jira Server')}
-            gotoParabol={goToParabol}
-            meetingRef={meeting}
-          />
-        </TabContents>
-        <TabContents>
-          <ScopePhaseAreaGitLab
-            isActive={isTabActive('GitLab')}
-            gotoParabol={goToParabol}
-            meetingRef={meeting}
-          />
-        </TabContents>
+        {/* swipeable views won't ignore null children: https://github.com/oliviertassinari/react-swipeable-views/issues/271 */}
+        {tabs.map(({label, Component}) => (
+          <TabContents key={label}>
+            <Component
+              meetingRef={meeting}
+              isActive={isTabActive(label)}
+              gotoParabol={gotoParabol}
+            />
+          </TabContents>
+        ))}
       </SwipeableViews>
     </ScopingArea>
   )
@@ -192,6 +192,7 @@ export default createFragmentContainer(ScopePhaseArea, {
       ...ScopePhaseAreaJira_meeting
       ...ScopePhaseAreaJiraServer_meeting
       ...ScopePhaseAreaParabolScoping_meeting
+      ...ScopePhaseAreaAzureDevOps_meeting
       endedAt
       localPhase {
         ...ScopePhaseArea_phase @relay(mask: false)
@@ -219,11 +220,17 @@ export default createFragmentContainer(ScopePhaseArea, {
                 id
               }
             }
+            azureDevOps {
+              sharedProviders {
+                id
+              }
+            }
           }
         }
         user {
           featureFlags {
             gitlab
+            azureDevOps
           }
         }
       }