+++ packages/client/components/ViewerNotOnTeam.tsx
@@ -1,27 +1,45 @@
-import {ViewerNotOnTeam_viewer} from '../__generated__/ViewerNotOnTeam_viewer.graphql'
-import React, {useEffect} from 'react'
-import {createFragmentContainer} from 'react-relay'
 import graphql from 'babel-plugin-relay/macro'
-import Ellipsis from './Ellipsis/Ellipsis'
-import PrimaryButton from './PrimaryButton'
-import AcceptTeamInvitationMutation from '../mutations/AcceptTeamInvitationMutation'
+import React, {useEffect} from 'react'
+import {PreloadedQuery, usePreloadedQuery} from 'react-relay'
 import useAtmosphere from '../hooks/useAtmosphere'
-import InvitationCenteredCopy from './InvitationCenteredCopy'
-import InviteDialog from './InviteDialog'
-import DialogContent from './DialogContent'
-import InvitationDialogCopy from './InvitationDialogCopy'
-import DialogTitle from './DialogTitle'
-import TeamInvitationWrapper from './TeamInvitationWrapper'
-import useRouter from '../hooks/useRouter'
-import PushInvitationMutation from '../mutations/PushInvitationMutation'
 import useDocumentTitle from '../hooks/useDocumentTitle'
+import useRouter from '../hooks/useRouter'
+import AcceptTeamInvitationMutation from '../mutations/AcceptTeamInvitationMutation'
+import PushInvitationMutation from '../mutations/PushInvitationMutation'
+import {ViewerNotOnTeamQuery} from '../__generated__/ViewerNotOnTeamQuery.graphql'
+import DialogContent from './DialogContent'
+import DialogTitle from './DialogTitle'
+import Ellipsis from './Ellipsis/Ellipsis'
+import InvitationCenteredCopy from './InvitationCenteredCopy'
+import InvitationDialogCopy from './InvitationDialogCopy'
+import InviteDialog from './InviteDialog'
+import PrimaryButton from './PrimaryButton'
+import TeamInvitationWrapper from './TeamInvitationWrapper'
 
 interface Props {
-  viewer: ViewerNotOnTeam_viewer
+  queryRef: PreloadedQuery<ViewerNotOnTeamQuery>
 }
 
+const query = graphql`
+  query ViewerNotOnTeamQuery($teamId: ID, $meetingId: ID) {
+    viewer {
+      teamInvitation(teamId: $teamId, meetingId: $meetingId) {
+        teamInvitation {
+          token
+        }
+        teamId
+        meetingId
+      }
+    }
+  }
+`
+
 const ViewerNotOnTeam = (props: Props) => {
-  const {viewer} = props
+  const {queryRef} = props
+  const data = usePreloadedQuery<ViewerNotOnTeamQuery>(query, queryRef, {
+    UNSTABLE_renderPolicy: 'full'
+  })
+  const {viewer} = data
   const {
     teamInvitation: {teamInvitation, meetingId, teamId}
   } = viewer
@@ -67,16 +85,4 @@ const ViewerNotOnTeam = (props: Props) => {
   )
 }
 
-export default createFragmentContainer(ViewerNotOnTeam, {
-  viewer: graphql`
-    fragment ViewerNotOnTeam_viewer on User {
-      teamInvitation(teamId: $teamId, meetingId: $meetingId) {
-        teamInvitation {
-          token
-        }
-        teamId
-        meetingId
-      }
-    }
-  `
-})
+export default ViewerNotOnTeam