+++ packages/client/modules/outcomeCard/components/OutcomeCardAssignMenu/TaskFooterUserAssigneeMenu.tsx
@@ -1,8 +1,10 @@
-import {TaskFooterUserAssigneeMenu_task} from '../../../../__generated__/TaskFooterUserAssigneeMenu_task.graphql'
-import {TaskFooterUserAssigneeMenu_viewer} from '../../../../__generated__/TaskFooterUserAssigneeMenu_viewer.graphql'
-import React, {useMemo} from 'react'
-import {createFragmentContainer} from 'react-relay'
 import graphql from 'babel-plugin-relay/macro'
+import React, {useMemo} from 'react'
+import {createFragmentContainer, PreloadedQuery, usePreloadedQuery} from 'react-relay'
+import {EmptyDropdownMenuItemLabel} from '~/components/EmptyDropdownMenuItemLabel'
+import {SearchMenuItem} from '~/components/SearchMenuItem'
+import useSearchFilter from '~/hooks/useSearchFilter'
+import {AreaEnum} from '~/__generated__/UpdateTaskMutation.graphql'
 import DropdownMenuLabel from '../../../../components/DropdownMenuLabel'
 import Menu from '../../../../components/Menu'
 import MenuAvatar from '../../../../components/MenuAvatar'
@@ -12,19 +14,40 @@ import useAtmosphere from '../../../../hooks/useAtmosphere'
 import {MenuProps} from '../../../../hooks/useMenu'
 import UpdateTaskMutation from '../../../../mutations/UpdateTaskMutation'
 import avatarUser from '../../../../styles/theme/images/avatar-user.svg'
-import {AreaEnum} from '~/__generated__/UpdateTaskMutation.graphql'
-import useSearchFilter from '~/hooks/useSearchFilter'
-import {SearchMenuItem} from '~/components/SearchMenuItem'
-import {EmptyDropdownMenuItemLabel} from '~/components/EmptyDropdownMenuItemLabel'
+import {TaskFooterUserAssigneeMenuQuery} from '../../../../__generated__/TaskFooterUserAssigneeMenuQuery.graphql'
+import {TaskFooterUserAssigneeMenu_task} from '../../../../__generated__/TaskFooterUserAssigneeMenu_task.graphql'
+
 interface Props {
   area: AreaEnum
   menuProps: MenuProps
-  viewer: TaskFooterUserAssigneeMenu_viewer
+  queryRef: PreloadedQuery<TaskFooterUserAssigneeMenuQuery>
   task: TaskFooterUserAssigneeMenu_task
 }
 
+const gqlQuery = graphql`
+  query TaskFooterUserAssigneeMenuQuery($teamId: ID!) {
+    viewer {
+      id
+      team(teamId: $teamId) {
+        teamId: id
+        teamMembers(sortBy: "preferredName") {
+          id
+          picture
+          preferredName
+          userId
+        }
+      }
+    }
+  }
+`
+
 const TaskFooterUserAssigneeMenu = (props: Props) => {
-  const {area, menuProps, task, viewer} = props
+  const {area, menuProps, task, queryRef} = props
+  const data = usePreloadedQuery<TaskFooterUserAssigneeMenuQuery>(gqlQuery, queryRef, {
+    UNSTABLE_renderPolicy: 'full'
+  })
+  const {viewer} = data
+
   const {userId, id: taskId} = task
   const {team} = viewer
   const atmosphere = useAtmosphere()
@@ -84,20 +107,6 @@ const TaskFooterUserAssigneeMenu = (props: Props) => {
 }
 
 export default createFragmentContainer(TaskFooterUserAssigneeMenu, {
-  viewer: graphql`
-    fragment TaskFooterUserAssigneeMenu_viewer on User {
-      id
-      team(teamId: $teamId) {
-        teamId: id
-        teamMembers(sortBy: "preferredName") {
-          id
-          picture
-          preferredName
-          userId
-        }
-      }
-    }
-  `,
   task: graphql`
     fragment TaskFooterUserAssigneeMenu_task on Task {
       id