+++ CHANGELOG.md
@@ -5,6 +5,63 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 This CHANGELOG follows conventions [outlined here](http://keepachangelog.com/).
 
+## 6.57.0 2022-May-4
+
+### Added
+ - feat(standups): upsertTeamPromptResponse (#6333)
+ - feat(ado): Initial Azure DevOps integration (#6260)
+ - feat(jira-server): Allow searching for issues in Sprint Poker (#6406)
+ - feat(sprint-poker): Can vote on GitLab issue (#6398)
+ - feat(sprint-poker): Push task to GitLab (#6427)
+
+### Changed
+ - chore: Create manualTestingGuide.md (#6426)
+ - chore: removed unused video related components (#6497)
+
+### Fixed
+ - fix(rethinkdb): attempt to fix the table rename (#6480)
+
+## 6.56.0 2022-April-27
+
+### Added
+
+- feat(sprint-poker): Track GitLab events (#6367)
+- feat(standups): Responses grid with static prompt (#6353)
+- feat(standups): Response Cards (#6392)
+- feat(standups): Discussion Drawer (#6370)
+- feat(CI): use prod build for integration tests (#6379)
+- feat(lint): Lint client (#6335)
+- feat(DX): Fast dev mode (#6337)
+- feat: Update illustration of empty discussion threads (#6423)
+
+### Changed
+
+- docs: Update CONTRIBUTING (#6432)
+- chore(deps): bump nconf from 0.11.3 to 0.11.4 (#6438)
+- refactor: update renderQuery to Suspense + Relay Hooks #5297 (#6251)
+
+### Fixed
+
+- fix: viewerMeetingMember can be undefined (#6441)
+- fix: ignore comparison order for equality (#6411)
+
+## 6.55.0 2022-April-20
+
+### Added
+
+- feat(jira-server): sprint poker vote to comment (#6341)
+- feat(sprint-poker): Add GitLab issue (#6267)
+- feat(sprint-poker): Search GitLab issues (#6290)
+- feat(sprint-poker): New Poker Scope UI (#6344)
+- feat(standups): Options menu - end meeting (#6342)
+- feat: Display a message when there are no more items to paginate (#6338)
+- chore: Track integrations in createTask (#6332)
+- chore(dx): Don't truncate TypeScript types (#6387)
+
+### Changed
+
+- build: update sharp library (#6383)
+
 ## 6.54.0 2022-April-13
 
 ### Fixed